Delete old offerbook and markets tab
