+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/MuSigController.java
@@ -20,10 +20,7 @@ package bisq.desktop.main.content.mu_sig;
 import bisq.desktop.ServiceProvider;
 import bisq.desktop.common.view.Controller;
 import bisq.desktop.main.content.ContentTabController;
-import bisq.desktop.main.content.mu_sig.market.MuSigMarketTabController;
 import bisq.desktop.main.content.mu_sig.offerbook.MuSigOfferbookController;
-import bisq.desktop.main.content.mu_sig.old_offerbook.buy.MuSigBuyOfferbookTabController;
-import bisq.desktop.main.content.mu_sig.old_offerbook.sell.MuSigSellOfferbookTabController;
 import bisq.desktop.main.content.mu_sig.portfolio.MuSigPortfolioTabController;
 import bisq.desktop.navigation.NavigationTarget;
 import lombok.Getter;
@@ -43,11 +40,7 @@ public class MuSigController extends ContentTabController<MuSigModel> {
     @Override
     protected Optional<? extends Controller> createController(NavigationTarget navigationTarget) {
         return switch (navigationTarget) {
-           // case MU_SIG_ONBOARDING -> Optional.of(new MuSigOnboardingController(serviceProvider));
             case MU_SIG_OFFERBOOK -> Optional.of(new MuSigOfferbookController(serviceProvider));
-            case MU_SIG_MARKET -> Optional.of(new MuSigMarketTabController(serviceProvider));
-            case MU_SIG_OFFERBOOK_BUY -> Optional.of(new MuSigBuyOfferbookTabController(serviceProvider));
-            case MU_SIG_OFFERBOOK_SELL -> Optional.of(new MuSigSellOfferbookTabController(serviceProvider));
             case MU_SIG_PORTFOLIO -> Optional.of(new MuSigPortfolioTabController(serviceProvider));
             default -> Optional.empty();
         };
