+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/MuSigModel.java
@@ -25,6 +25,6 @@ import lombok.Getter;
 public class MuSigModel extends ContentTabModel {
     @Override
     public NavigationTarget getDefaultNavigationTarget() {
-        return NavigationTarget.MU_SIG_OFFERBOOK_BUY;
+        return NavigationTarget.MU_SIG_OFFERBOOK;
     }
 }
