--- apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/market/chart/MuSigMarketChartController.java
@@ -1,43 +1,0 @@
-/*
- * This file is part of Bisq.
- *
- * Bisq is free software: you can redistribute it and/or modify it
- * under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or (at
- * your option) any later version.
- *
- * Bisq is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package bisq.desktop.main.content.mu_sig.market.chart;
-
-import bisq.desktop.ServiceProvider;
-import bisq.desktop.common.view.Controller;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class MuSigMarketChartController implements Controller {
-    @Getter
-    private final MuSigMarketChartView view;
-    private final MuSigMarketChartModel model;
-
-    public MuSigMarketChartController(ServiceProvider serviceProvider) {
-        model = new MuSigMarketChartModel();
-        view = new MuSigMarketChartView(model, this);
-    }
-
-    @Override
-    public void onActivate() {
-    }
-
-    @Override
-    public void onDeactivate() {
-    }
-}
