--- apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/old_offerbook/crypto_btc/xmr/MuSigOfferbookXmrModel.java
@@ -1,37 +1,0 @@
-/*
- * This file is part of Bisq.
- *
- * Bisq is free software: you can redistribute it and/or modify it
- * under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or (at
- * your option) any later version.
- *
- * Bisq is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package bisq.desktop.main.content.mu_sig.old_offerbook.crypto_btc.xmr;
-
-import bisq.common.currency.Market;
-import bisq.desktop.main.content.mu_sig.old_offerbook.crypto_btc.MuSigOfferbookCryptoBtcModel;
-import bisq.offer.Direction;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
-
-import java.util.Collections;
-
-@Slf4j
-@Getter
-public class MuSigOfferbookXmrModel extends MuSigOfferbookCryptoBtcModel {
-    public MuSigOfferbookXmrModel(Direction direction, Market xmrMarket) {
-        super(direction);
-        getSelectedMarket().set(xmrMarket);
-        setMarketPredicate(item -> item.getOffer().getMarket().equals(xmrMarket));
-        getMarkets().setAll(Collections.singletonList(xmrMarket));
-    }
-}
