+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/portfolio/MuSigLevel2TabModel.java
@@ -15,16 +15,13 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.mu_sig.old_offerbook.crypto_btc;
+package bisq.desktop.main.content.mu_sig.portfolio;
 
-import bisq.desktop.ServiceProvider;
-import bisq.desktop.main.content.mu_sig.old_offerbook.MuSigOfferbookController;
-import bisq.offer.Direction;
+import bisq.desktop.common.view.TabModel;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public abstract class MuSigOfferbookCryptoBtcController<M extends MuSigOfferbookCryptoBtcModel, V extends MuSigOfferbookCryptoBtcView<?, ?>> extends MuSigOfferbookController<M, V> {
-    public MuSigOfferbookCryptoBtcController(ServiceProvider serviceProvider, Direction direction) {
-        super(serviceProvider, direction);
-    }
+@Getter
+public abstract class MuSigLevel2TabModel extends TabModel {
 }
