+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/mu_sig/portfolio/MuSigPortfolioTabView.java
@@ -18,7 +18,6 @@
 package bisq.desktop.main.content.mu_sig.portfolio;
 
 import bisq.desktop.common.Styles;
-import bisq.desktop.main.content.mu_sig.old_offerbook.MuSigLevel2TabView;
 import bisq.desktop.navigation.NavigationTarget;
 import bisq.i18n.Res;
 import lombok.extern.slf4j.Slf4j;
