+++ apps/desktop/desktop/src/main/java/bisq/desktop/navigation/NavigationTarget.java
@@ -149,19 +149,7 @@ public enum NavigationTarget {
 
     // MU_SIG
     MU_SIG(CONTENT),
-    MU_SIG_MARKET(MU_SIG),
     MU_SIG_OFFERBOOK(MU_SIG),
-    MU_SIG_MARKET_CHART(MU_SIG_MARKET),
-    MU_SIG_MARKET_BY_CURRENCY(MU_SIG_MARKET),
-    MU_SIG_MARKET_BY_PAYMENT_METHODS(MU_SIG_MARKET),
-    MU_SIG_OFFERBOOK_BUY(MU_SIG),
-    MU_SIG_OFFERBOOK_BUY_BTC(MU_SIG_OFFERBOOK_BUY),
-    MU_SIG_OFFERBOOK_BUY_XMR(MU_SIG_OFFERBOOK_BUY),
-    MU_SIG_OFFERBOOK_BUY_OTHER(MU_SIG_OFFERBOOK_BUY),
-    MU_SIG_OFFERBOOK_SELL(MU_SIG),
-    MU_SIG_OFFERBOOK_SELL_BTC(MU_SIG_OFFERBOOK_SELL),
-    MU_SIG_OFFERBOOK_SELL_XMR(MU_SIG_OFFERBOOK_SELL),
-    MU_SIG_OFFERBOOK_SELL_OTHER(MU_SIG_OFFERBOOK_SELL),
     MU_SIG_PORTFOLIO(MU_SIG),
     MU_SIG_OPEN_OFFERS(MU_SIG_PORTFOLIO),
     MU_SIG_OPEN_TRADES(MU_SIG_PORTFOLIO),
