+++ settings/src/main/java/bisq/settings/CookieKey.java
@@ -49,8 +49,6 @@ public enum CookieKey {
     BISQ_EASY_OFFER_LIST_PAYMENT_FILTERS(true),
     BISQ_EASY_OFFER_LIST_CUSTOM_PAYMENT_FILTER(true),
     CREATE_MU_SIG_OFFER_IS_MIN_AMOUNT_ENABLED,
-    MU_SIG_OFFERBOOK_SELECTED_BTC_MARKET,
-    MU_SIG_OFFERBOOK_SELECTED_OTHER_MARKET,
     MU_SIG_MARKETS_FILTER,
     MU_SIG_MARKET_SORT_TYPE,
     MU_SIG_OFFER_DIRECTION_FILTER;
